Rework to get both stil and hw to compile Added new code to driver_temp to properly process raw values Moved away from custom macro access functions Using smaller datatypes where possible Added various new x86 sitl drivers Switched to custom printf replacement "str", which uses absolute types No longer using uint32_t

+++ hc3d-tm/src/business_logic/data_reader.h
@@ -12,7 +12,7 @@
 #include "stdbool.h"
 #include "config.h"
 
-extern int16_t data_reader_last_temperatures[HC3D_CONFIG_TEMP_BUF_SIZE][HC3D_CONFIG_TEMP_SENSOR_COUNT];
+extern uint8_t data_reader_last_temperatures[HC3D_CONFIG_TEMP_BUF_SIZE][HC3D_CONFIG_TEMP_SENSOR_COUNT];
 
 // Returns true when all sensors are settled, false otherwise
 // Reads out all sensors and update the moving buffer found
